% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info.R
\name{check_info}
\alias{check_info}
\title{check_info function}
\usage{
check_info()
}
\value{
check_info.RDS
}
\description{
This function allows you to save information on dqm check process
}
\examples{
check_info()
}
\keyword{check}
