% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_DB.R
\name{connect_DB}
\alias{connect_DB}
\title{connect_DB function}
\usage{
connect_DB()
}
\value{
con object
}
\description{
This function allows you to connect R and CDM DB
}
\examples{
connect_DB()
}
\keyword{connect}
