% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete.R
\name{is_complete}
\alias{is_complete}
\title{is_complete function}
\usage{
is_complete(data)
}
\value{
list of result and error messages
}
\description{
This function allows you to bring the result of completeness check
}
\examples{
is_complete(x)
}
\keyword{completeness}
