% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_consistent.R
\name{is_consistent}
\alias{is_consistent}
\title{is_consistent function}
\usage{
is_consistent(data)
}
\value{
list of result and error messages
}
\description{
This function allows you to bring the result of consistency check
}
\examples{
is_consistent(x)
}
\keyword{consistency}
