% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unique}
\alias{is_unique}
\title{is_unique function}
\usage{
is_unique(data)
}
\value{
list of result and error messages
}
\description{
This function allows you to bring the result of uniqueness check
}
\examples{
is_unique(x)
}
\keyword{uniqueness}
