% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_accuracy.R
\name{moa_accuracy}
\alias{moa_accuracy}
\title{moa_accuracy function}
\usage{
moa_accuracy()
}
\value{
csv file
}
\description{
This function allows you to perform accuracy check
}
\examples{
moa_accuracy(x)
}
\keyword{accuracy}
