% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_completeness.R
\name{moa_completeness}
\alias{moa_completeness}
\title{moa_completeness function}
\usage{
moa_completeness()
}
\value{
csv file
}
\description{
This function allows you to perform completeness check
}
\examples{
moa_completeness(x)
}
\keyword{completeness}
