% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_consistency.R
\name{moa_consistency}
\alias{moa_consistency}
\title{moa_consistency function}
\usage{
moa_consistency()
}
\value{
csv file
}
\description{
This function allows you to perform consistency check
}
\examples{
moa_consistency(x)
}
\keyword{consistency}
