% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_initialize.R
\name{moa_initialize}
\alias{moa_initialize}
\title{moa_initialize function}
\usage{
moa_initialize()
}
\value{
rds file
}
\description{
This function allows you to initialize dqm process
}
\examples{
moa_initialize()
}
\keyword{initialize}
