% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_lv2.R
\name{moa_lv2}
\alias{moa_lv2}
\title{moa_lv2 function}
\usage{
moa_lv2(
  input_schema = NULL,
  input_table = NULL,
  input_contents = NULL,
  input_group = NULL,
  concept_id = NULL
)
}
\value{
rds file
}
\description{
This function allows you to create a summary table and a graph for lv 2 dqm analysis
}
\examples{
moa_lv2(schema, table, contents, group, concept_id)
}
\keyword{analysis}
\keyword{lv2}
