% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_tablerowindex.R
\name{moa_tablerowindex}
\alias{moa_tablerowindex}
\title{moa_tablerowindex function}
\usage{
moa_tablerowindex(con)
}
\value{
rds file
}
\description{
This function allows you to write table row index
}
\examples{
moa_tablerowindex(x)
}
\keyword{tablerowindex}
