% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_validity.R
\name{moa_validity}
\alias{moa_validity}
\title{moa_validity function}
\usage{
moa_validity(con)
}
\value{
csv file
}
\description{
This function allows you to perform validity check
}
\examples{
moa_validity(x)
}
\keyword{validity}
