% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_overview.R
\name{update_overview}
\alias{update_overview}
\title{update_overview function}
\usage{
update_overview()
}
\value{
dashboard_overview, no_rules
}
\description{
This function allows you to update overview score table
}
\examples{
update_overview()
}
\keyword{overview}
